giftcards/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-filter==24.3
django-paypal==2.1
djangorestframework==3.15.2
fonttools==4.55.3
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
Naked==0.1.32
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pycryptodome==3.22.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
six==1.17.0
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
